Move Point, ColorMode, RGBA out
