+++ SDLWrapper/ColorMode.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "include.h"
+#include "begin_code.h"
+class ColorMode
+{
+public:
+    int r, g, b;
+    ColorMode(int R, int G, int B);
+    ColorMode();
+};
+#include "end_code.h"
